+++ node/src/types/kernel/api/block/updateBlock/index.d.ts
@@ -17,13 +17,7 @@
 
 
 /* directories */
-export * as appendBlock from "./appendBlock";
-export * as getBlockKramdown from "./getBlockKramdown";
-export * as getChildBlocks from "./getChildBlocks";
-export * as insertBlock from "./insertBlock";
-export * as prependBlock from "./prependBlock";
-export * as updateBlock from "./updateBlock";
 
 /* files */
-export * as getBlockBreadcrumb from "./getBlockBreadcrumb";
-export * as getDocInfo from "./getDocInfo";
+export * from "./payload";
+export * from "./response";